assure that protocol::Context::to_url() never shows passwords (#450)
